src/test/java/com/spartaglobal/ymao/sorters/InsertionSortTest.java
package com.spartaglobal.ymao.sorters;

import com.spartaglobal.ymao.model.sorters.InsertionSort;
import org.junit.Test;

import static org.junit.Assert.assertArrayEquals;

public class InsertionSortTest extends SortTest {

    @Override
    @Test
    public void testLarge() {
        initialiseLarge();
        assertArrayEquals(getResult(),
                InsertionSort.getInstance().sort(getInput()));
    }

    @Override
    @Test
    public void testZero() {
        initialiseZero();
        assertArrayEquals(getResult(),
                InsertionSort.getInstance().sort(getInput()));
    }

    @Override
    @Test
    public void testSorted() {
        initialiseLarge();
        assertArrayEquals(getResult(),
                InsertionSort.getInstance().sort(getResult()));
    }

}
